+++ soundpool_platform_interface/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 2.1.0
+- Platform-specific options support
 ## 2.0.0-nullsafety.0
 - Null-safety support
 ## 1.0.1